+++ src/django_dto_field/features/base.py
@@ -9,6 +9,7 @@ class DtoCodeEnum(bytes, Enum):
     """Enum for builtin DTO codes."""
 
     DICT = b"\x01"
+    DATACLASS = b"\x02"
 
 
 class BaseDtoFeature(ABC):
@@ -16,9 +17,14 @@ class BaseDtoFeature(ABC):
 
     dto_code: ClassVar[bytes | None] = None
 
-    def _get_dto_code(self) -> bytes:
-        if self.dto_code is None:
+    def __init__(self, *args, **kwargs) -> None:
+        super().__init__(*args, **kwargs)
+        self._get_dto_code()  # <-- calling just to be sure that dto_code is set.
+
+    @classmethod
+    def _get_dto_code(cls) -> bytes:
+        if cls.dto_code is None:
             raise DtoFeatureError(
                 "Dto Feature Error: missing definition of `dto_code`."
             )
-        return self.dto_code
+        return cls.dto_code